Move full page logic to FullPageScroller component
